+++ tests/scripts/test_deploy_dag.py
@@ -90,6 +90,13 @@ def copy_config_files_and_set_tmp_folder_names_as_ids(
         )
     )
     generate_dag.write_to_file(pipeline_yaml_str, pipeline_path / "pipeline.yaml")
+    (ENV_DATASETS_PATH / dataset_path.name / pipeline_path.name).mkdir(
+        parents=True, exist_ok=True
+    )
+    shutil.copyfile(
+        pipeline_path / "pipeline.yaml",
+        ENV_DATASETS_PATH / dataset_path.name / pipeline_path.name / "pipeline.yaml",
+    )
 
 
 def create_airflow_folders(airflow_home: pathlib.Path):
@@ -175,7 +182,7 @@ def test_script_always_requires_dataset_arg(
 pipeline_path_2 = pipeline_path
 
 
-def test_script_can_deploy_without_variables_file(
+def test_script_can_deploy_without_variables_files(
     dataset_path: pathlib.Path,
     pipeline_path: pathlib.Path,
     airflow_home: pathlib.Path,
@@ -190,13 +197,21 @@ def test_script_can_deploy_without_variables_file(
         f"{dataset_path.name}_variables.json",
     )
 
-    # Delete the variables file
-    vars_file = f"{dataset_path.name}_variables.json"
-    (ENV_DATASETS_PATH / dataset_path.name / vars_file).unlink()
-    assert not (ENV_DATASETS_PATH / dataset_path.name / vars_file).exists()
+    # Delete the shared variables file
+    (ENV_DATASETS_PATH / "shared_variables.json").unlink()
+    assert not (ENV_DATASETS_PATH / "shared_variables.json").exists()
+
+    # Delete the dataset-specific variables file
+    (
+        ENV_DATASETS_PATH / dataset_path.name / f"{dataset_path.name}_variables.json"
+    ).unlink()
+    assert not (
+        ENV_DATASETS_PATH / dataset_path.name / f"{dataset_path.name}_variables.json"
+    ).exists()
 
     mocker.patch("scripts.deploy_dag.run_gsutil_cmd")
     mocker.patch("scripts.deploy_dag.run_cloud_composer_vars_import")
+    mocker.patch("scripts.deploy_dag.composer_airflow_version", return_value=2)
 
     deploy_dag.main(
         local=False,
@@ -208,6 +223,37 @@ def test_script_can_deploy_without_variables_file(
         composer_bucket="test-bucket",
         composer_region="test-region",
     )
+
+
+def test_script_errors_out_when_deploying_airflow2_dag_to_airflow1_env(
+    dataset_path: pathlib.Path,
+    pipeline_path: pathlib.Path,
+    airflow_home: pathlib.Path,
+    env: str,
+    mocker,
+):
+    setup_dag_and_variables(
+        dataset_path,
+        pipeline_path,
+        airflow_home,
+        env,
+        f"{dataset_path.name}_variables.json",
+    )
+
+    mocker.patch("scripts.deploy_dag.get_dag_airflow_version", return_value=2)
+    mocker.patch("scripts.deploy_dag.composer_airflow_version", return_value=1)
+
+    with pytest.raises(Exception):
+        deploy_dag.main(
+            local=False,
+            env_path=ENV_PATH,
+            dataset_id=dataset_path.name,
+            pipeline=pipeline_path.name,
+            airflow_home=airflow_home,
+            composer_env="test-env",
+            composer_bucket="test-bucket",
+            composer_region="test-region",
+        )
 
 
 def test_script_with_pipeline_arg_deploys_only_that_pipeline(
@@ -382,6 +428,7 @@ def test_script_with_local_flag_copies_files_to_local_airflow_env(
         composer_region=None,
     )
 
+    assert (airflow_home / "data" / "variables" / "shared_variables.json").exists()
     assert (airflow_home / "data" / "variables" / variables_filename).exists()
     assert (airflow_home / "dags" / dag_filename).exists()